+++ src/Diagnostics/DiagnosticsResponseAnalyzer.ts
@@ -1,9 +1,9 @@
-import type { server } from "typescript/lib/tsserverlibrary";
-import { TSServer } from "../TSServer";
-import { TestWorkspace } from "../Workspaces/TestWorkspace";
-import { CodeAction } from "./Actions/CodeAction";
-import { Diagnostic } from "./Diagnostic";
-import { ICodeAnalysisResult } from "./ICodeAnalysisResult";
+import type { server } from "typescript/lib/tsserverlibrary.js";
+import { TSServer } from "../TSServer.js";
+import { TestWorkspace } from "../Workspaces/TestWorkspace.js";
+import { CodeAction } from "./Actions/CodeAction.js";
+import { Diagnostic } from "./Diagnostic.js";
+import { ICodeAnalysisResult } from "./ICodeAnalysisResult.js";
 
 /**
  * Provides the functionality to analyze a diagnostic-response.